mcp_agent/requirements.txt
# 核心LangGraph和LangChain依赖
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# MCP相关依赖
langchain-mcp-adapters>=0.1.0
mcp>=1.0.0

# FastAPI服务器依赖
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# HTTP客户端依赖
aiohttp>=3.9.0
httpx>=0.25.0

# 其他工具依赖
python-dotenv>=1.0.0
typing-extensions>=4.8.0

# 其他依赖
langgraph-api>=0.2.3
langgraph-sdk>=0.1.61
langchain-community>=0.3.0
asyncio
python-multipart>=0.0.6
sqlite3
uuid
datetime

# 数据库相关依赖
sqlalchemy>=2.0.0
aiomysql>=0.2.0
pymysql>=1.1.0

# Redis相关依赖
redis>=5.0.0
aioredis>=2.0.0

# 其他工具
cryptography>=41.0.0 